Quaternion cleaning.
- moves code from header to .inl
- use typedef to express the inner types instead of specifying them all the time
  The inner types are: Quat, Vec3, Mat3x3, Real
- add Quaternion to forward declaration file.
- remove the use of the implicit auto converting function (BREAKING)
template<Real2> void toMatrix(Matrix<3,3,Real2>){}
As there is no more hidden type conversion the caller's have to handle themselves
the type conversion before calling instead of relying on the inner-conversion. This is
good as it propagate to the user-level that there is type conversion involved instead of
hidden that fact. The consequence is that it is not possible to do stuff like that anymore
Quat<float>::toMatrix(Mat<3,3,double)>